traffic density/local_requirements.txt
ultralytics>=8.0.0
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.7.0
numpy>=1.23.0
pandas>=1.5.0
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.6.0
